reuse context menu logic
