+++ apps/chat/src/components/Chat/TalkTo/ItemCardView.tsx
@@ -37,7 +37,7 @@ import { ReplayAsIsIcon } from '@/src/components/Chat/ReplayAsIsIcon';
 import { ModelIcon } from '@/src/components/Chatbar/ModelIcon';
 import { EntityMarkdownDescription } from '@/src/components/Common/MarkdownDescription';
 import { ShareIcon } from '@/src/components/Common/ShareIcon';
-import { MarketplaceEntityContextMenuRenderer } from '@/src/components/Marketplace/EntityContextMenu/MarketplaceEntityContextMenuRenderer';
+import { MarketplaceEntityContextMenu } from '@/src/components/Marketplace/EntityContextMenu/MarketplaceEntityContextMenu';
 import { MarketplaceEntityIndicator } from '@/src/components/Marketplace/MarketplaceEntityIndicator';
 import { TopicsList } from '@/src/components/Marketplace/TopicsList';
 
@@ -143,7 +143,7 @@ export const ItemCardView = <T extends MarketplaceEntity>({
     >
       {hasContextMenu && (
         <div className="absolute right-4 top-4 flex cursor-pointer gap-1 xl:right-5 xl:top-5">
-          <MarketplaceEntityContextMenuRenderer
+          <MarketplaceEntityContextMenu
             className="xl:invisible group-hover:xl:visible"
             entity={entity}
             disabledActions={
